refactor(RacingGame): 게임을 진행하는 역할을 컨트롤러에서 분리한다
