+++ src/test/java/racingcar/RacingGameTest.java
@@ -8,12 +8,12 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CarControllerTest {
+public class RacingGameTest {
 
     @Test
     void 우승자를_선출한다_우승자가_한명인경우() {
         // given
-        CarController carController = new CarController();
+        RacingGame racingGame = new RacingGame();
         Car car1 = new Car();
         Car car2 = new Car();
         List<Car> cars = new ArrayList<>();
@@ -26,13 +26,13 @@ public class CarControllerTest {
         car2.attemptToMove(3);
 
         // then
-        assertThat(carController.findWinners(cars)).containsExactly(car1);
+        assertThat(racingGame.findWinners(cars)).containsExactly(car1);
     }
 
     @Test
     void 우승자를_선출한다_우승자가_두명인경우() {
         // given
-        CarController carController = new CarController();
+        RacingGame racingGame = new RacingGame();
         Car car1 = new Car();
         Car car2 = new Car();
 
@@ -46,7 +46,7 @@ public class CarControllerTest {
         car2.attemptToMove(4);
 
         // then
-        assertThat(carController.findWinners(cars)).containsExactly(car1, car2);
+        assertThat(racingGame.findWinners(cars)).containsExactly(car1, car2);
     }
